Update user_story_issue_template.md
Update the template to remove unused examples to make it easier for contributors to fill in information, instead of seeing the actual example from the template. The examples can still be seen, but they are now comments.

Added the original idea to add the ability to convert tasks into separate issues.